init navigation
