+++ ui/components/Logo.tsx
@@ -16,7 +16,7 @@ type Props = {
 function Logo({ className, link = V2Routes.Automations, collapsed }: Props) {
   const dark = useInDarkMode();
   return (
-    <Flex className={className} wide>
+    <Flex className={className} alignItems="center">
       <Link to={link}>
         <img src={dark ? images.logoDark : images.logoLight} />
       </Link>
@@ -42,7 +42,8 @@ export default styled(Logo)`
     height: 32px;
   }
   padding-left: 20px;
+  height: 60px;
   //nav width: 200px - space btwn nav and content: 24px - content padding: 24px. All together 248px - 20 for left padding to line up with detail page titles.
-  width: ${(props) => (props.collapsed ? "92px" : "228px")};
+  width: ${(props) => (props.collapsed ? "36px" : "180px")};
   transition: width 0.5s;
 `;
